@@ -1,19 +1,18 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <Version>0.5.0</Version>
+    <Version>0.6.0</Version>
     <PackAsTool>true</PackAsTool>
     <ToolCommandName>act2txt</ToolCommandName>
     <PackageOutputPath>nupkg</PackageOutputPath>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <PackageReleaseNotes><![CDATA[
-- Restructured code    
-- Added unit tests
-- Fixed a bug in parsing the transparent color index in ACT extra data 
-(which unit tests caught)
+- Target .NET 8.0
+- Make sure StructLayout is correct
+- Make building easier
 ]]></PackageReleaseNotes>
   </PropertyGroup>
 
